requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
bokeh==0.12.13
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
colorama==0.3.9
ConfigArgParse==0.12.0
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
dateparser==0.6.0
decorator==4.2.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
int-date==0.1.8
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.10
MarkupSafe==1.0
matplotlib==2.1.2
numpy==1.14.0
pandas==0.22.0
pandas-datareader==0.5.0
parso==0.1.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-binance==0.6.2
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
regex==2018.1.10
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
ruamel.yaml==0.15.35
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
stockstats==0.2.0
tornado==4.5.3
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
tzlocal==1.5.1
urllib3==1.22
wcwidth==0.1.7
zope.interface==4.4.3
